+++ src/Subscriber/CacheEventsSubscriber.php
@@ -3,11 +3,13 @@
 namespace DigaShopwareCacheHelper\Subscriber;
 
 use Psr\Log\LoggerInterface;
-use Symfony\Component\EventDispatcher\EventSubscriberInterface;
-use Shopware\Core\Framework\Adapter\Cache\InvalidateCacheEvent;
-use Shopware\Storefront\Framework\Cache\Event\HttpCacheHitEvent;
-use Shopware\Storefront\Framework\Cache\Event\HttpCacheItemWrittenEvent;
+use Shopware\Recovery\Common\HttpClient\Response;
 use Shopware\Core\System\SystemConfig\SystemConfigService;
+use Shopware\Core\Framework\Adapter\Cache\InvalidateCacheEvent;
+use Symfony\Component\EventDispatcher\EventSubscriberInterface;
+use Shopware\Storefront\Framework\Cache\Event\HttpCacheHitEvent;
+use Shopware\Storefront\Framework\Cache\Event\HttpCacheGenerateKeyEvent;
+use Shopware\Storefront\Framework\Cache\Event\HttpCacheItemWrittenEvent;
 
 class CacheEventsSubscriber implements EventSubscriberInterface
 {
@@ -32,7 +34,8 @@ class CacheEventsSubscriber implements EventSubscriberInterface
         return [
             HttpCacheItemWrittenEvent::class => 'onCacheItemWritten',
             HttpCacheHitEvent::class => 'onCacheHit',
-            InvalidateCacheEvent::class => 'onInvalidateCache'
+            InvalidateCacheEvent::class => 'onInvalidateCache',
+            HttpCacheGenerateKeyEvent::class => 'onHttpCacheGenerateKeyEvent'
         ];
     }
 
@@ -45,7 +48,12 @@ class CacheEventsSubscriber implements EventSubscriberInterface
             if($logOnCacheHit){
                 $requestUri = $event->getRequest()->getRequestUri();
                 $itemKey = $event->getItem()->getKey();        
-                $this->logger->info('CacheHitEvent ItemKey: ' . $itemKey . ' RequestUri: '. $requestUri);
+
+                /** @var Response $response */                
+                $ttl = $event->getResponse()->getTtl();
+                $maxAge = $event->getResponse()->getMaxAge();
+
+                $this->logger->info('CacheHitEvent ItemKey: ' . $itemKey . ' TTL: ' .  $ttl . ' maxAge: ' .  $maxAge . ' RequestUri: '. $requestUri);
             }
             
         } catch (\Throwable $th) {       
@@ -57,12 +65,22 @@ class CacheEventsSubscriber implements EventSubscriberInterface
     {
         try {
             $logOnCacheItemWritten = $this->systemConfigService->get('DigaShopwareCacheHelper.config.logOnCacheItemWritten');
+            $logTagsOnCacheItemWritten = $this->systemConfigService->get('DigaShopwareCacheHelper.config.logTagsOnCacheItemWritten');
 
             if($logOnCacheItemWritten){
                 $requestUri = $event->getRequest()->getRequestUri();
                 $itemKey = $event->getItem()->getKey();        
                 $tags = $event->getTags();
-                $this->logger->info('CacheItemWrittenEvent ItemKey: ' . $itemKey . ' Tags: ' .  json_encode($tags) . ' RequestUri: '. $requestUri);
+    
+                /** @var Response $response */                
+                $ttl = $event->getResponse()->getTtl();
+                $maxAge = $event->getResponse()->getMaxAge();
+                
+                if($logTagsOnCacheItemWritten){
+                    $this->logger->info('CacheItemWrittenEvent ItemKey: ' . $itemKey . ' Tags: ' .  json_encode($tags) . ' TTL: ' .  $ttl . ' maxAge: ' .  $maxAge . ' RequestUri: '. $requestUri);
+                }else{
+                    $this->logger->info('CacheItemWrittenEvent ItemKey: ' . $itemKey . ' TTL: ' .  $ttl . ' maxAge: ' .  $maxAge . ' RequestUri: '. $requestUri);
+                }                
             }
 
         } catch (\Throwable $th) {
@@ -81,6 +99,23 @@ class CacheEventsSubscriber implements EventSubscriberInterface
             }
 
         } catch (\Throwable $th) {
+            $this->logger->error( $th->getMessage());
+        }
+    }
+
+    public function onHttpCacheGenerateKeyEvent(HttpCacheGenerateKeyEvent $event): void
+    {
+        try {
+                         
+            $logHttpCacheGenerateKeyEvent = $this->systemConfigService->get('DigaShopwareCacheHelper.config.logHttpCacheGenerateKeyEvent');
+
+            if($logHttpCacheGenerateKeyEvent){
+                $requestUri = $event->getRequest()->getRequestUri();
+                $hash = $event->getHash();        
+                $this->logger->info('HttpCacheGenerateKeyEvent hash: ' . $hash . ' RequestUri: '. $requestUri);
+            }
+            
+        } catch (\Throwable $th) {       
             $this->logger->error( $th->getMessage());
         }
     }